init commit: Working
Essentially the goal of this is to, on an x86 host, point
run.sh at a clone of the linux-kernel. Then we cross compile the
kernel (and rpms) for ppc64le. All build artifacts go in a different
directory so we don't scuff up our clone.

We then implement a test.sh that starts qemu to run this kernel.
We can use `expect` to script some fun tests. And we can load
an external module for further quick probing into kernel space.

For now the external module is testing one of my branches
vnic_kasan_read_overflow.

Signed-off-by: Nick Child <nick.child@ibm.com>